[Native Protocol] Read and count operation  (#589)

+++ src/main/java/io/stargate/sgv2/jsonapi/service/operation/model/impl/CountOperationPage.java
@@ -5,7 +5,7 @@ import io.stargate.sgv2.jsonapi.api.model.command.CommandStatus;
 import java.util.Map;
 import java.util.function.Supplier;
 
-public record CountOperationPage(int count) implements Supplier<CommandResult> {
+public record CountOperationPage(long count) implements Supplier<CommandResult> {
   @Override
   public CommandResult get() {
     return new CommandResult(Map.of(CommandStatus.COUNTED_DOCUMENT, count()));
